feat: restructure main
