+++ src/main/patches/app-settings.ts
@@ -1,0 +1,12 @@
+Object.defineProperty(global, "appSettings", {
+  set: (v /* : typeof global.appSettings*/) => {
+    // cspell:ignore youre
+    v.set("DANGEROUS_ENABLE_DEVTOOLS_ONLY_ENABLE_IF_YOU_KNOW_WHAT_YOURE_DOING", true);
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-expect-error
+    delete global.appSettings;
+    global.appSettings = v;
+  },
+  get: () => global.appSettings,
+  configurable: true,
+});
