+++ src/main/patches/on-ready/permission-handler.ts
@@ -1,0 +1,14 @@
+import { session } from "electron";
+
+const defaultPermissionRequestHandler = session.defaultSession.setPermissionRequestHandler.bind(
+  session.defaultSession,
+);
+session.defaultSession.setPermissionRequestHandler = (cb) => {
+  defaultPermissionRequestHandler((webContents, permission, callback, details) => {
+    if (permission === "media") {
+      callback(true);
+      return;
+    }
+    cb?.(webContents, permission, callback, details);
+  });
+};
